(refactor) Remove custom overlay implementation
This PR removes the remnants of the custom overlay implementation which has since been replaced with workspaces. Changes include:

- Removing the custom overlay component system (hook, component, styles)
- Updating all related imports and references to use the workspace system
- Cleaning up the codebase to remove any remaining references to the overlay system

Other changes include renaming the stock operations dialog component to stock operations modal for consistency.
Overall, these changes improve consistency with the framework's workspace pattern and removes the custom overlay implementation in favour of the standard workspace system.
