+++ src/dashboard/home-dashboard.component.tsx
@@ -3,7 +3,6 @@ import { useParams } from 'react-router-dom';
 import { useLayoutType, isDesktop, useExtensionStore, ExtensionSlot, WorkspaceContainer } from '@openmrs/esm-framework';
 import { type DashboardConfig } from '../types';
 import DashboardView from './dashboard-view.component';
-import Overlay from '../core/components/overlay/overlay.component';
 import styles from './home-dashboard.scss';
 
 export default function Dashboard() {
@@ -18,15 +17,12 @@ export default function Dashboard() {
   const activeDashboard = dashboards.find((dashboard) => dashboard.name === params?.dashboard) || dashboards[0];
 
   return (
-    <>
-      <Overlay />
-      <div className={styles.homePageWrapper}>
-        <section className={isDesktop(layout) ? styles.dashboardContainer : styles.dashboardContainerTablet}>
-          {isDesktop(layout) && <ExtensionSlot name="stock-sidebar-slot" key={layout} />}
-          <DashboardView title={activeDashboard?.name} dashboardSlot={activeDashboard?.slot} />
-        </section>
-        <WorkspaceContainer overlay contextKey="stock-management" />
-      </div>
-    </>
+    <div className={styles.homePageWrapper}>
+      <section className={isDesktop(layout) ? styles.dashboardContainer : styles.dashboardContainerTablet}>
+        {isDesktop(layout) && <ExtensionSlot name="stock-sidebar-slot" key={layout} />}
+        <DashboardView title={activeDashboard?.name} dashboardSlot={activeDashboard?.slot} />
+      </section>
+      <WorkspaceContainer overlay contextKey="stock-management" />
+    </div>
   );
 }
