+++ src/stock-operations/stock-operations-forms/step2.test.tsx
@@ -48,10 +48,6 @@ jest.mock('../stock-operations.resource', () => ({
   }),
 }));
 
-jest.mock('../../core/components/overlay/hook', () => ({
-  closeOverlay: jest.fn(),
-}));
-
 jest.mock('../../stock-items/stock-items.resource', () => ({
   useStockItem: jest.fn().mockReturnValue({
     isLoading: false,
