+++ src/stock-operations/stock-operations-modal/stock-operations-print-button.component.tsx
