+++ src/stock-operations/stock-operations-modal/stock-operations.scss
