+++ translations/en.json
@@ -327,7 +327,6 @@
   "submittingForReview": "Submitting for review",
   "success": "Success",
   "successfullysaved": "You have successfully saved user role scope",
-  "tabletOverlay": "Tablet overlay",
   "tags": "Tags",
   "timeTaken": "Time Taken",
   "to": "To",
